integrate session

+++ pidev_web/src/Controller/SesssionsFormController.php
@@ -11,6 +11,7 @@ use App\Entity\Courses;
 use App\Entity\Participations;
 use App\Services\PaymentService;
 use App\Services\CalendarService;
+use Illuminate\Support\Facades\Redirect;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\HttpFoundation\JsonResponse;
@@ -18,6 +19,7 @@ use Symfony\Component\HttpFoundation\RedirectResponse;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\Routing\Annotation\Route;
 use Symfony\Component\Serializer\Normalizer\NormalizerInterface;
+use Symfony\Component\Security\Core\Security;
 
 if (!isset($_SESSION)) {
   session_start();
@@ -29,17 +31,21 @@ class SesssionsFormController extends AbstractController
   private EntityManagerInterface $entityManager;
   private PaymentService $paymentService;
   private CalendarService $calendarService;
-  private int $tmpUserId;
+  private ?int $userId = null;
 
 
-  public function __construct(SessionsRepository $sessionsRepository, EntityManagerInterface $entityManager)
+  public function __construct(SessionsRepository $sessionsRepository, EntityManagerInterface $entityManager, Security $security)
   {
     $this->entityManager = $entityManager;
     $this->sessionsRepository = $sessionsRepository;
     $this->paymentService = new PaymentService();
     $this->calendarService = new CalendarService();
-    $this->tmpUserId = $_ENV['TMP_USER_ID'];
+
+    $user = $security->getUser();
+    if ($user)
+      $this->userId = $security->getUser()->getId();
   }
+
 
   private function buildSession(array $inputs, int $userId, string $imageLink, array $resources, EntityManagerInterface $entityManager): Sessions
   {
@@ -85,7 +91,7 @@ class SesssionsFormController extends AbstractController
 
   private function buildSessionObject(int $id): Sessions
   {
-    $session = $this->sessionsRepository->getPageSession($id, $this->tmpUserId);
+    $session = $this->sessionsRepository->getPageSession($id, $this->userId);
     $creatorId = $session->getUser()->getId();
 
 
@@ -133,6 +139,9 @@ class SesssionsFormController extends AbstractController
   public function persistSession(Request $request)
   {
 
+    if ($this->userId == null)
+      return $this->redirectToRoute('app_login');
+
     $imageFileLink = '';
     $imageFile = $_FILES['image'];
     if ($imageFile['error'] != 4)
@@ -145,7 +154,7 @@ class SesssionsFormController extends AbstractController
     }
 
 
-    $session = $this->buildSession($_POST, $this->tmpUserId, $imageFileLink, $resourceFiles,  $this->entityManager);
+    $session = $this->buildSession($_POST, $this->userId, $imageFileLink, $resourceFiles,  $this->entityManager);
     $error = $this->validate($session);
     if ($error) {
       $this->entityManager->clear();
@@ -155,13 +164,13 @@ class SesssionsFormController extends AbstractController
     $this->entityManager->persist($session);
     $this->entityManager->flush();
 
-    if (!is_dir("uploads/" . $this->tmpUserId))
-      mkdir("uploads/" . $this->tmpUserId);
+    if (!is_dir("uploads/" . $this->userId))
+      mkdir("uploads/" . $this->userId);
 
-    move_uploaded_file($imageFile['tmp_name'], "uploads/" . $this->tmpUserId . "/" . $imageFileLink);
+    move_uploaded_file($imageFile['tmp_name'], "uploads/" . $this->userId . "/" . $imageFileLink);
 
     foreach ($resourceFiles as $resourceFile)
-      move_uploaded_file($resourceFile[2], "uploads/" . $this->tmpUserId . "/" . $resourceFile[0]);
+      move_uploaded_file($resourceFile[2], "uploads/" . $this->userId . "/" . $resourceFile[0]);
 
     if (!empty($_POST['id']))
       return $this->redirect($request->query->get('redirect') . '?success=1');
@@ -172,8 +181,12 @@ class SesssionsFormController extends AbstractController
   }
 
   #[Route('/removeSession/{id}', name: 'sessions_list_remove_session')]
-  public function removeSession(Sessions $session): Response
+  public function removeSession(int $id): Response
   {
+    if ($this->userId == null)
+      return $this->redirectToRoute('app_login');
+
+    $session = $this->sessionsRepository->find($id);
     $this->entityManager->remove($session);
     $this->entityManager->flush();
     return $this->redirectToRoute('sessions_list', ['success' => '1']);
@@ -182,6 +195,10 @@ class SesssionsFormController extends AbstractController
   #[Route('/getSessionForm/{id}', name: 'get_session_form')]
   public function getSessionForm(NormalizerInterface $normalizer, int $id): JsonResponse
   {
+
+    if ($this->userId == null)
+      return $this->redirectToRoute('app_login');
+
     $session = $this->buildSessionObject($id);
     $normalizedSession = $normalizer->normalize($session, 'json', ["groups" => "sessionFrom"]);
     return new JsonResponse($normalizedSession);
@@ -190,6 +207,9 @@ class SesssionsFormController extends AbstractController
   #[Route('/admin/blockSession/{id}', name: 'block_session', methods: ['PUT'])]
   public function blockSession(int $id): JsonResponse
   {
+    if ($this->userId == null)
+      return $this->redirectToRoute('app_login');
+
     $this->sessionsRepository->setSessionBlock($id, true);
     return new JsonResponse(['error' => null]);
   }
@@ -197,6 +217,9 @@ class SesssionsFormController extends AbstractController
   #[Route('/admin/unblockSession/{id}', name: 'unblock_session', methods: ['PUT'])]
   public function unblockSession(int $id): JsonResponse
   {
+    if ($this->userId == null)
+      return $this->redirectToRoute('app_login');
+
     $this->sessionsRepository->setSessionBlock($id, false);
     return new JsonResponse(['error' => null]);
   }
@@ -204,10 +227,14 @@ class SesssionsFormController extends AbstractController
   #[Route('/participate/{id}', name: 'participate_session')]
   public function participateSession(int $id): RedirectResponse
   {
-    $session = $this->sessionsRepository->getParticipationSession($id);
-    $user = $this->entityManager->getRepository(Users::class)->find($this->tmpUserId);
 
-    $participation = $this->entityManager->getRepository(Participations::class)->getParticipation($id, $this->tmpUserId);
+    if ($this->userId == null)
+      return $this->redirectToRoute('app_login');
+
+    $session = $this->sessionsRepository->getParticipationSession($id);
+    $user = $this->entityManager->getRepository(Users::class)->find($this->userId);
+
+    $participation = $this->entityManager->getRepository(Participations::class)->getParticipation($id, $this->userId);
 
     if ($participation == null) {
       $participation = new Participations();
@@ -231,6 +258,10 @@ class SesssionsFormController extends AbstractController
   #[Route('/validateParticipation', name: 'validate_participation')]
   public function validateParticipation(Request $request): Response
   {
+
+    if ($this->userId == null)
+      return $this->redirectToRoute('app_login');
+
     $paymentRef = $request->query->get('payment_ref');
     $payment = $this->paymentService->getPayment($paymentRef);
     $participationId = $payment['orderId'];
@@ -259,6 +290,10 @@ class SesssionsFormController extends AbstractController
   #[Route('/createEvent', name: 'create_event')]
   public function createEvent(): Response
   {
+
+    if ($this->userId == null)
+      return $this->redirectToRoute('app_login');
+
     if ($_SESSION['calendarOp'] == 'addAttendee') {
       $url = $this->calendarService->authUser('addAttendee');
       if ($url != null)
